requirements.txt
flask
flask-cors
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pillow
tensorboard==2.13.0
gast==0.4.0
protobuf==3.20.3
typing-extensions==4.5.0